update_fake_score_every + remove_droput fixed

+++ sid_train.py
@@ -107,7 +107,7 @@ class CommaSeparatedList(click.ParamType):
 # Parameters for IDMD
 @click.option('--aug_dim',    help='', metavar='STR', type=str, default="inf", show_default=True)
 @click.option('--loss_sid_or_idmd',    help='', metavar='STR', type=str, default="sid", show_default=True)
-@click.option('--update_fake_score_iters',    help='', metavar='FLOAT', type=float, default=1, show_default=True)
+@click.option('--update_fake_score_iters',    help='', metavar='INT', type=int, default=1, show_default=True)
 @click.option('--remove_dropout',    help='', metavar='BOOL', type=bool, default=False, show_default=True)
